pyproject.toml
[project]
name = "ecomagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "db-dtypes==1.2.0",
    "google-cloud-bigquery>=3.13.0",
    "ipython>=9.4.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=1.0.0",
    "langgraph>=0.2.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "graphviz>=0.21",
    "langchain-community>=0.3.28",
]
